requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
spotipy==2.23.0
gunicorn==20.1.0
Flask-CORS==3.0.10
Flask-Migrate==4.0.4